requirements.txt
boto3==1.26.40
botocore==1.29.40
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.13
